move compute Otsu's threshold function from core to the dswx_hls module
